refactor(api): reformat api, add jupyter to devspace tools
